code/requirements.txt
beautifulsoup4==4.3.2
cffi==0.9.2
cryptography==0.8.2
cssselect==0.9.1
enum34==1.0.4
lxml==3.4.4
nltk==3.0.2
pyasn1==0.1.7
pycparser==2.12
pyOpenSSL==0.15.1
queuelib==1.2.2
requests==2.7.0
Scrapy==0.24.6
six==1.9.0
SQLAlchemy==1.0.2
Twisted==15.1.0
w3lib==1.11.0
zope.interface==4.1.2
Flask==0.10.1
Flask-WTF==0.9.5
Flask-Bootstrap==3.2.0.2
Jinja2==2.7.1
WTForms==1.0.5
Werkzeug==0.9.4
Pillow==2.8.1


